Only put in StyledComponents.js if used by multiple components
